Pull a reusable autocompleting search filter input from the SearchFilter component.
